#20 - loading character data from db, added sample character for testing

+++ src/main/java/model/Character.java
@@ -8,7 +8,7 @@ public class Character extends PanacheEntity {
 //    public int playerId;
     public String name;
 //    public int raceId;
-    public int classId;
+    public long classId;
     public int backgroundId;
     public int level;
     public boolean isCreated;
